requirements.txt
numpy
opencv-python
pandas
pillow
pywin32
pygetwindow
pytk
easyocr